+++ include/stumpless/target/buffer.h
@@ -26,12 +26,15 @@
 extern "C" {
 #  endif
 
-  void stumpless_close_buffer_target( struct stumpless_target *target );
-  struct stumpless_target *stumpless_open_buffer_target( const char *name,
-                                                         char *buffer,
-                                                         size_t size,
-                                                         int options,
-                                                         int default_facility );
+  void
+  stumpless_close_buffer_target( struct stumpless_target *target );
+
+  struct stumpless_target *
+  stumpless_open_buffer_target( const char *name,
+                                char *buffer,
+                                size_t size,
+                                int options,
+                                int default_facility );
 
 #  ifdef __cplusplus
 }                               /* extern "C" */