+++ src/target/socket.c
@@ -16,6 +16,7 @@
  * limitations under the License.
  */
 
+#include <stdio.h>
 #include <stddef.h>
 #include <string.h>
 #include <unistd.h>
@@ -28,6 +29,8 @@
 #include "private/memory.h"
 #include "private/target.h"
 #include "private/target/socket.h"
+
+static size_t next_socket_number = 0;
 
 void
 stumpless_close_socket_target( struct stumpless_target *target ) {
@@ -45,11 +48,14 @@ stumpless_close_socket_target( struct stumpless_target *target ) {
 }
 
 struct stumpless_target *
-stumpless_open_socket_target( const char *name, int options,
+stumpless_open_socket_target( const char *name,
+                              const char *local_socket,
+                              int options,
                               int default_facility ) {
   struct stumpless_target *pub_target;
   struct socket_target *priv_target;
-  size_t name_len;
+  size_t name_len, local_socket_len;
+  char default_socket[15];
 
   clear_error(  );
 
@@ -60,18 +66,50 @@ stumpless_open_socket_target( const char *name, int options,
 
   pub_target = alloc_mem( sizeof( *pub_target ) );
   if( !pub_target ) {
+    printf("pub target allocation failed\n");
     goto fail;
   }
 
   name_len = strlen( name );
-  priv_target = new_socket_target( name, name_len );
+
+  if( !local_socket ) {
+    default_socket[0] = 's';
+    default_socket[1] = 't';
+    default_socket[2] = 'u';
+    default_socket[3] = 'm';
+    default_socket[4] = 'p';
+    default_socket[5] = 'l';
+    default_socket[6] = 'e';
+    default_socket[7] = 's';
+    default_socket[8] = 's';
+    default_socket[9] = '-';
+    default_socket[10] = ( ( next_socket_number >> 12 ) & 0xf ) + 97;
+    default_socket[11] = ( ( next_socket_number >> 8 ) & 0xf ) + 97;
+    default_socket[12] = ( ( next_socket_number >> 4 ) & 0xf ) + 97;
+    default_socket[13] = ( next_socket_number & 0xf ) + 97;
+    default_socket[14] = '\0';
+    next_socket_number++;
+    priv_target = new_socket_target( name, name_len, default_socket, 15 );
+  } else { 
+    local_socket_len = strlen( local_socket );
+    priv_target = new_socket_target( name, name_len, local_socket, local_socket_len );
+  }
+
   if( !priv_target ) {
+    printf("private target creation failed\n");
     goto fail_priv_target;
   }
 
-  pub_target->name = alloc_mem( name_len );
+  pub_target->name = alloc_mem( name_len + 1 );
   if( !pub_target->name ) {
+    printf("name allocation failed\n");
     goto fail_pub_name;
+  }
+
+  pub_target->id = register_priv_target( priv_target );
+  if( pub_target->id < 0 ) {
+    printf("id creation failed\n");
+    goto fail_id;
   }
 
   memcpy( pub_target->name, name, name_len );
@@ -80,13 +118,14 @@ stumpless_open_socket_target( const char *name, int options,
   pub_target->options = options;
   pub_target->default_prival =
     get_prival( default_facility, STUMPLESS_SEVERITY_INFO );
-  pub_target->id = register_priv_target( priv_target );
 
   stumpless_set_current_target( pub_target );
   return pub_target;
 
+fail_id:
+  free_mem( pub_target->name );
 fail_pub_name:
-  free_mem( priv_target );
+  destroy_socket_target( priv_target );
 fail_priv_target:
   free_mem( pub_target );
 fail:
@@ -103,30 +142,33 @@ destroy_socket_target( struct socket_target *trgt ) {
   }
 
   close( trgt->local_socket );
-  unlink( "stmplss-tst" );
+  unlink( trgt->local_addr.sun_path );
   free_mem( trgt );
 }
 
 struct socket_target *
-new_socket_target( const char *dest, size_t dest_len ) {
+new_socket_target( const char *dest, size_t dest_len,
+                   const char *source, size_t source_len ) {
   struct socket_target *trgt;
 
   trgt = alloc_mem( sizeof( *trgt ) );
   if( !trgt ) {
+    printf("could not allocate private target memory\n");
     return NULL;
   }
 
   trgt->target_addr.sun_family = AF_UNIX;
-  // todo need to check dest_len before this memcpy happens
   memcpy( &trgt->target_addr.sun_path, dest, dest_len );
   trgt->target_addr.sun_path[dest_len] = '\0';
 
   trgt->local_addr.sun_family = AF_UNIX;
-  memcpy( &trgt->local_addr.sun_path, "stmplss-tst", 12 );
+  memcpy( &trgt->local_addr.sun_path, source, source_len );
+  trgt->local_addr.sun_path[source_len] = '\0';
 
   trgt->local_socket = socket( trgt->local_addr.sun_family, SOCK_DGRAM, 0 );
   if( trgt->local_socket < 0 ) {
     free_mem( trgt );
+    printf("could not open local socket\n");
     return NULL;
   }
 
@@ -134,6 +176,7 @@ new_socket_target( const char *dest, size_t dest_len ) {
       ( trgt->local_socket, ( struct sockaddr * ) &trgt->local_addr,
         sizeof( trgt->local_addr ) ) < 0 ) {
     free_mem( trgt );
+    printf("could not bind to local socket\n");
     return NULL;
   }