+++ ChangeLog.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
  - Add file target capability.
+ - Add windows event log target capability.
 
-## [0.1.1] - 2018-07-21
+## [1.0.0] - 2018-XX-XX
 ### Added
  - Add `stumpless_set_realloc` function.
+
+### Changed
+ - Added local socket parameter to `stumpless_open_socket_target` function.
 
 ### Fixed
  - Targets are now removed and their memory freed upon closure.